vader#run: improve collecting: use explicit files, throw error
This patch allows to call `:Vader foo.vader.new`, which gets not collected
with `:Vader foo*` still.

In case no test cases are found an exception/error is thrown, which
helps when calling it from a wrong cwd etc: since stdout is redirected
usually, Vim would appear to just hang without this patch.