Set g:vader_case_ok before After blocks
`After` blocks get called always, but it is useful for them to know if
the test succeeded so far.

I.e. with Neomake I am checking if windows have been closed when tearing
down the tests, but this should be skipped if the test itself did not
finish because of a previous failure (and therefore windows are likely
not closed).